+++ components/camel-dropbox/src/test/java/org/apache/camel/component/dropbox/integration/producer/DropboxProducerGetFolderIT.java
@@ -17,6 +17,7 @@
 package org.apache.camel.component.dropbox.integration.producer;
 
 import java.io.IOException;
+import java.util.Map;
 
 import org.apache.camel.builder.RouteBuilder;
 import org.apache.camel.component.dropbox.integration.DropboxTestSupport;
@@ -25,15 +26,22 @@ import org.apache.camel.component.dropbox.util.DropboxResultHeader;
 import org.apache.camel.component.mock.MockEndpoint;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.condition.EnabledIf;
 
-public class DropboxProducerSearchQueryTest extends DropboxTestSupport {
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
-    public static final String FILE_NAME = "myTestFile.txt";
+@EnabledIf("org.apache.camel.component.dropbox.integration.DropboxTestSupport#hasCredentials")
+public class DropboxProducerGetFolderIT extends DropboxTestSupport {
+
+    public static final String FILE_NAME1 = "myFile.txt";
+    public static final String FILE_NAME2 = "myFile2.txt";
+    private static final String CONTENT1 = "content1";
+    private static final String CONTENT2 = "content2";
 
     @BeforeEach
     public void createFile() throws IOException {
-        final String content = "Hi camels";
-        createFile(FILE_NAME, content);
+        createFile(FILE_NAME1, CONTENT1);
+        createFile(FILE_NAME2, CONTENT2);
     }
 
     @Test
@@ -46,13 +54,24 @@ public class DropboxProducerSearchQueryTest extends DropboxTestSupport {
         test("direct:start2");
     }
 
+    @Test
+    public void testCamelDropboxHeaderHasPriorityOnParameter() throws Exception {
+        test("direct:start3");
+    }
+
     private void test(String endpoint) throws InterruptedException {
         template.sendBody(endpoint, null);
-
         MockEndpoint mock = getMockEndpoint("mock:result");
         mock.expectedMinimumMessageCount(1);
-        mock.message(0).header(DropboxResultHeader.FOUND_FILES.name()).contains(String.format("%s/%s", workdir, FILE_NAME));
+        mock.message(0).header(DropboxResultHeader.DOWNLOADED_FILES.name())
+                .contains(String.format("%s/%s", workdir, FILE_NAME1));
+        mock.message(0).header(DropboxResultHeader.DOWNLOADED_FILES.name())
+                .contains(String.format("%s/%s", workdir, FILE_NAME2));
         mock.assertIsSatisfied();
+
+        final Map<String, byte[]> items = mock.getExchanges().get(0).getIn().getBody(Map.class);
+        assertEquals(CONTENT1, new String(items.get(String.format("%s/%s", workdir, FILE_NAME1))));
+        assertEquals(CONTENT2, new String(items.get(String.format("%s/%s", workdir, FILE_NAME2))));
     }
 
     @Override
@@ -60,15 +79,19 @@ public class DropboxProducerSearchQueryTest extends DropboxTestSupport {
         return new RouteBuilder() {
             public void configure() {
                 from("direct:start")
-                        .to(String.format("dropbox://search?accessToken={{accessToken}}&remotePath=%s&query=%s", workdir,
-                                FILE_NAME))
+                        .to("dropbox://get?accessToken={{accessToken}}&remotePath=" + workdir)
                         .to("mock:result");
 
                 from("direct:start2")
                         .setHeader(DropboxConstants.HEADER_REMOTE_PATH, constant(workdir))
-                        .setHeader(DropboxConstants.HEADER_QUERY, constant(FILE_NAME))
-                        .to("dropbox://search?accessToken={{accessToken}}")
+                        .to("dropbox://get?accessToken={{accessToken}}")
                         .to("mock:result");
+
+                from("direct:start3")
+                        .setHeader(DropboxConstants.HEADER_REMOTE_PATH, constant(workdir))
+                        .to("dropbox://get?accessToken={{accessToken}}&remotePath=/aWrongPath")
+                        .to("mock:result");
+
             }
         };
     }
