+++ components/camel-dropbox/src/test/java/org/apache/camel/component/dropbox/integration/producer/DropboxProducerSearchQueryIT.java
@@ -17,7 +17,6 @@
 package org.apache.camel.component.dropbox.integration.producer;
 
 import java.io.IOException;
-import java.util.Map;
 
 import org.apache.camel.builder.RouteBuilder;
 import org.apache.camel.component.dropbox.integration.DropboxTestSupport;
@@ -26,20 +25,17 @@ import org.apache.camel.component.dropbox.util.DropboxResultHeader;
 import org.apache.camel.component.mock.MockEndpoint;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.condition.EnabledIf;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
+@EnabledIf("org.apache.camel.component.dropbox.integration.DropboxTestSupport#hasCredentials")
+public class DropboxProducerSearchQueryIT extends DropboxTestSupport {
 
-public class DropboxProducerGetFolderTest extends DropboxTestSupport {
-
-    public static final String FILE_NAME1 = "myFile.txt";
-    public static final String FILE_NAME2 = "myFile2.txt";
-    private static final String CONTENT1 = "content1";
-    private static final String CONTENT2 = "content2";
+    public static final String FILE_NAME = "myTestFile.txt";
 
     @BeforeEach
     public void createFile() throws IOException {
-        createFile(FILE_NAME1, CONTENT1);
-        createFile(FILE_NAME2, CONTENT2);
+        final String content = "Hi camels";
+        createFile(FILE_NAME, content);
     }
 
     @Test
@@ -52,24 +48,13 @@ public class DropboxProducerGetFolderTest extends DropboxTestSupport {
         test("direct:start2");
     }
 
-    @Test
-    public void testCamelDropboxHeaderHasPriorityOnParameter() throws Exception {
-        test("direct:start3");
-    }
-
     private void test(String endpoint) throws InterruptedException {
         template.sendBody(endpoint, null);
+
         MockEndpoint mock = getMockEndpoint("mock:result");
         mock.expectedMinimumMessageCount(1);
-        mock.message(0).header(DropboxResultHeader.DOWNLOADED_FILES.name())
-                .contains(String.format("%s/%s", workdir, FILE_NAME1));
-        mock.message(0).header(DropboxResultHeader.DOWNLOADED_FILES.name())
-                .contains(String.format("%s/%s", workdir, FILE_NAME2));
+        mock.message(0).header(DropboxResultHeader.FOUND_FILES.name()).contains(String.format("%s/%s", workdir, FILE_NAME));
         mock.assertIsSatisfied();
-
-        final Map<String, byte[]> items = mock.getExchanges().get(0).getIn().getBody(Map.class);
-        assertEquals(CONTENT1, new String(items.get(String.format("%s/%s", workdir, FILE_NAME1))));
-        assertEquals(CONTENT2, new String(items.get(String.format("%s/%s", workdir, FILE_NAME2))));
     }
 
     @Override
@@ -77,19 +62,15 @@ public class DropboxProducerGetFolderTest extends DropboxTestSupport {
         return new RouteBuilder() {
             public void configure() {
                 from("direct:start")
-                        .to("dropbox://get?accessToken={{accessToken}}&remotePath=" + workdir)
+                        .to(String.format("dropbox://search?accessToken={{accessToken}}&remotePath=%s&query=%s", workdir,
+                                FILE_NAME))
                         .to("mock:result");
 
                 from("direct:start2")
                         .setHeader(DropboxConstants.HEADER_REMOTE_PATH, constant(workdir))
-                        .to("dropbox://get?accessToken={{accessToken}}")
+                        .setHeader(DropboxConstants.HEADER_QUERY, constant(FILE_NAME))
+                        .to("dropbox://search?accessToken={{accessToken}}")
                         .to("mock:result");
-
-                from("direct:start3")
-                        .setHeader(DropboxConstants.HEADER_REMOTE_PATH, constant(workdir))
-                        .to("dropbox://get?accessToken={{accessToken}}&remotePath=/aWrongPath")
-                        .to("mock:result");
-
             }
         };
     }
