+++ components/camel-dropbox/src/test/resources/test-options.properties
@@ -15,6 +15,6 @@
 ## limitations under the License.
 ## ---------------------------------------------------------------------------
 
-accessToken=<api-token>
+accessToken=
 clientIdentifier=camel-dropbox
 workDir=/camel-test
